refactor: rename the distribution class names after code review
