+++ configs/dummy_pipelines/experimental_1.yaml
@@ -2,38 +2,38 @@ n_experiments: 2
 output_dir: "results_dummy/experimental"
 input_dir: "data/experimental"
 seed: 42
-#data_generation:
-#  experimental: True
-#  method: "experimental"
-#  n_samples: 20
-#  default_model_name: "humanTRB"
-#  input_columns: ["junction_aa", "v_call", "j_call"]
-#models:
-#    - name: "pwm"
-#      config: "configs/dummy_generative_models/PWM.yaml"
-#      train_dir: "train"
-#      test_dir: "test"
-#    - name: "sonnia"
-#      config: "configs/dummy_generative_models/soNNia.yaml"
-#      train_dir: "train"
-#      test_dir: "test"
-#    - name: "vae"
-#      config: "configs/dummy_generative_models/VAE.yaml"
-#      train_dir: "train"
-#      test_dir: "test"
+data_generation:
+  experimental: True
+  method: "experimental"
+  n_samples: 20
+  default_model_name: "humanTRB"
+  input_columns: ["junction_aa", "v_call", "j_call"]
+models:
+    - name: "pwm"
+      config: "configs/dummy_generative_models/PWM.yaml"
+      train_dir: "train"
+      test_dir: "test"
+    - name: "sonnia"
+      config: "configs/dummy_generative_models/soNNia.yaml"
+      train_dir: "train"
+      test_dir: "test"
+    - name: "vae"
+      config: "configs/dummy_generative_models/VAE.yaml"
+      train_dir: "train"
+      test_dir: "test"
 analyses:
-#    - name: "novelty"
-#      model_names: ["pwm", "sonnia"]
-#      default_model_name: "humanTRB"
-#    - name: "phenotype"
-#      model_names: ["pwm"]
-#      default_model_name: "humanTRB"
-#    - name: "phenotype"
-#      model_names: ["sonnia"]
-#      default_model_name: "humanTRB"
-#    - name: "phenotype"
-#      model_names: [ "vae" ]
-#      default_model_name: "humanTRB"
+    - name: "novelty"
+      model_names: ["pwm", "sonnia"]
+      default_model_name: "humanTRB"
+    - name: "phenotype"
+      model_names: ["pwm"]
+      default_model_name: "humanTRB"
+    - name: "phenotype"
+      model_names: ["sonnia"]
+      default_model_name: "humanTRB"
+    - name: "phenotype"
+      model_names: [ "vae" ]
+      default_model_name: "humanTRB"
     - name: "reduced_dimensionality"
       model_names: [ "pwm", "sonnia", "vae" ]
       default_model_name: "humanTRB"
@@ -42,30 +42,30 @@ analyses:
       model_names: [ "pwm", "sonnia", "vae" ]
       default_model_name: "humanTRB"
       reference_data: "test"
-#    - name: "length_distribution"
-#      model_names: [ "pwm", "sonnia", "vae" ]
-#      default_model_name: "humanTRB"
-#      reference_data: "train"
-#    - name: "length_distribution"
-#      model_names: [ "pwm", "sonnia", "vae" ]
-#      default_model_name: "humanTRB"
-#      reference_data: "test"
-#    - name: "kmer_distribution"
-#      model_names: [ "pwm", "sonnia", "vae" ]
-#      default_model_name: "humanTRB"
-#      reference_data: "train"
-#    - name: "kmer_distribution"
-#      model_names: [ "pwm", "sonnia", "vae" ]
-#      default_model_name: "humanTRB"
-#      reference_data: "test"
-#    - name: "pgen"
-#      model_names: [ "pwm", "sonnia", "vae" ]
-#      default_model_name: "humanTRB"
-#    - name: "aminoacid_distribution"
-#      model_names: [ "pwm", "sonnia", "vae" ]
-#      default_model_name: "humanTRB"
-#      reference_data: "train"
-#    - name: "aminoacid_distribution"
-#      model_names: [ "pwm", "sonnia", "vae" ]
-#      default_model_name: "humanTRB"
-#      reference_data: "test"
+    - name: "length_distribution"
+      model_names: [ "pwm", "sonnia", "vae" ]
+      default_model_name: "humanTRB"
+      reference_data: "train"
+    - name: "length_distribution"
+      model_names: [ "pwm", "sonnia", "vae" ]
+      default_model_name: "humanTRB"
+      reference_data: "test"
+    - name: "kmer_distribution"
+      model_names: [ "pwm", "sonnia", "vae" ]
+      default_model_name: "humanTRB"
+      reference_data: "train"
+    - name: "kmer_distribution"
+      model_names: [ "pwm", "sonnia", "vae" ]
+      default_model_name: "humanTRB"
+      reference_data: "test"
+    - name: "pgen"
+      model_names: [ "pwm", "sonnia", "vae" ]
+      default_model_name: "humanTRB"
+    - name: "aminoacid_distribution"
+      model_names: [ "pwm", "sonnia", "vae" ]
+      default_model_name: "humanTRB"
+      reference_data: "train"
+    - name: "aminoacid_distribution"
+      model_names: [ "pwm", "sonnia", "vae" ]
+      default_model_name: "humanTRB"
+      reference_data: "test"
