+++ gen_airr_bm/analysis/distribution/distribution_factory.py
@@ -1,11 +1,11 @@
-from gen_airr_bm.analysis.distribution.aa_strategy import AADistributionStrategy
-from gen_airr_bm.analysis.distribution.base_strategy import DistributionStrategy
-from gen_airr_bm.analysis.distribution.kmer_strategy import KmerDistributionStrategy
-from gen_airr_bm.analysis.distribution.length_strategy import LengthDistributionStrategy
+from gen_airr_bm.analysis.distribution.aa_distribution_strategy import AADistributionStrategy
+from gen_airr_bm.analysis.distribution.base_distribution_strategy import BaseDistributionStrategy
+from gen_airr_bm.analysis.distribution.kmer_distribution_strategy import KmerDistributionStrategy
+from gen_airr_bm.analysis.distribution.length_distribution_strategy import LengthDistributionStrategy
 from gen_airr_bm.constants.distribution_type import DistributionType
 
 
-def get_distribution_strategy(distribution_type: DistributionType, k: int = 3) -> DistributionStrategy:
+def get_distribution_strategy(distribution_type: DistributionType, k: int = 3) -> BaseDistributionStrategy:
     if distribution_type == DistributionType.AA:
         return AADistributionStrategy()
     elif distribution_type == DistributionType.KMER:
