+++ gen_airr_bm/analysis/distribution/length_distribution_strategy.py
@@ -2,10 +2,10 @@ from collections import Counter
 
 from scipy.spatial.distance import jensenshannon
 
-from gen_airr_bm.analysis.distribution.base_strategy import DistributionStrategy
+from gen_airr_bm.analysis.distribution.base_distribution_strategy import BaseDistributionStrategy
 
 
-class LengthDistributionStrategy(DistributionStrategy):
+class LengthDistributionStrategy(BaseDistributionStrategy):
     def compute_divergence(self, gen_seqs, ref_seqs):
         gen_lengths = [len(seq) for seq in gen_seqs]
         ref_lengths = [len(seq) for seq in ref_seqs]
