+++ tests/analysis/distribution/test_base_distribution_strategy.py
@@ -1,36 +1,36 @@
 import pytest
 import numpy as np
 
-from gen_airr_bm.analysis.distribution.base_strategy import DistributionStrategy
+from gen_airr_bm.analysis.distribution.base_distribution_strategy import BaseDistributionStrategy
 from gen_airr_bm.core.analysis_config import AnalysisConfig
 
 
 # Dummy subclass for testing concrete methods
-class DummyStrategy(DistributionStrategy):
+class DummyStrategyBase(BaseDistributionStrategy):
     def compute_divergence(self, gen_seqs, ref_seqs):
         return ["dummy"]
 
 
 def test_cannot_instantiate_abstract_class():
     with pytest.raises(TypeError):
-        DistributionStrategy()
+        BaseDistributionStrategy()
 
 
 def test_init_mean_std_scores():
-    strategy = DummyStrategy()
+    strategy = DummyStrategyBase()
     mean, std = strategy.init_mean_std_scores()
     assert mean == {}
     assert std == {}
 
 
 def test_init_divergence_scores():
-    strategy = DummyStrategy()
+    strategy = DummyStrategyBase()
     scores = strategy.init_divergence_scores()
     assert scores == []
 
 
 def test_update_divergence_scores():
-    strategy = DummyStrategy()
+    strategy = DummyStrategyBase()
     scores = [0.1, 0.2]
     new = [0.3, 0.4]
     strategy.update_divergence_scores(scores, new)
@@ -38,7 +38,7 @@ def test_update_divergence_scores():
 
 
 def test_update_mean_std_scores():
-    strategy = DummyStrategy()
+    strategy = DummyStrategyBase()
     mean_scores, std_scores = {}, {}
     divergence_scores = [0.5, 1.5, 2.5]
 
@@ -53,7 +53,7 @@ def test_update_mean_std_scores():
 def test_plot_scores_calls_plot_jsd_scores(mocker):
     mock_plot_jsd_scores = mocker.patch("gen_airr_bm.analysis.distribution.base_strategy.plot_jsd_scores")
 
-    strategy = DummyStrategy()
+    strategy = DummyStrategyBase()
     mean_scores = {"model": 0.5}
     std_scores = {"model": 0.1}
