+++ tests/analysis/distribution/test_kmer_distribution_strategy.py
@@ -1,5 +1,5 @@
 import pytest
-from gen_airr_bm.analysis.distribution.kmer_strategy import (
+from gen_airr_bm.analysis.distribution.kmer_distribution_strategy import (
     KmerDistributionStrategy,
     compute_kmer_distribution,
     compute_jsd_kmers
